app/requirements.txt
Django==2.2.8
django-rest-framework==0.1.0
djangorestframework==3.10.3
django-debug-toolbar==2.1
psycopg2-binary==2.8.3
flake8==3.7.5
pytz==2019.3
factory_boy==2.12.0
django-extensions==2.2.5
